adding tests and fixing selections comparing to VMD
