+++ moleculekit/atomselect/atomselect.py
@@ -1,538 +1,8 @@
-import moleculekit.ply.lex as lex
-import moleculekit.ply.yacc as yacc
-import numpy as np
-from moleculekit.util import sequenceID
-from moleculekit.molecule import Molecule
-
-# molecule types
-reserved = [
-    "protein",
-    "nucleic",
-    "lipid",
-    "lipids",
-    "ion",
-    "ions",
-    "water",
-    "waters",
-    "noh",
-    "hydrogen",
-    "backbone",
-    "sidechain",
-]
-# molecule properties
-reserved += [
-    "index",
-    "serial",
-    "name",
-    "element",
-    "resname",
-    "resid",
-    "residue",
-    "altloc",
-    "insertion",
-    "chain",
-    "segname",
-    "segid",
-    "mass",
-    "charge",
-    "occupancy",
-    "fragment",
-]
-# operands
-reserved += [
-    "and",
-    "or",
-    "within",
-    "exwithin",
-    "of",
-    "same",
-    "as",
-    "not",
-    "to",
-]
-# functions
-reserved += ["abs", "sqr"]
-# List of token names.   This is always required
-tokens = [
-    "PLUS",
-    "MINUS",
-    "TIMES",
-    "DIVIDE",
-    "LPAREN",
-    "RPAREN",
-    "ID",
-    "INTEGER",
-    "FLOAT",
-    "NORMSTRING",
-    "NORMSTRINGSINGLE",
-    "EQUAL",
-    "LESSER",
-    "GREATER",
-    "LESSEREQUAL",
-    "GREATEREQUAL",
-    "XCOOR",
-    "YCOOR",
-    "ZCOOR",
-] + [rr.upper() for rr in reserved]
-
-# Regular expression rules for simple tokens
-t_PLUS = r"\+"
-t_MINUS = r"-"
-t_TIMES = r"\*"
-t_DIVIDE = r"/"
-t_LPAREN = r"\("
-t_RPAREN = r"\)"
-t_EQUAL = r"\="
-t_LESSER = r"\<"
-t_GREATER = r"\>"
-t_LESSEREQUAL = r"\<\="
-t_GREATEREQUAL = r"\>\="
-
-
-def t_XCOOR(t):
-    r"x"
-    return t
-
-
-def t_YCOOR(t):
-    r"y"
-    return t
-
-
-def t_ZCOOR(t):
-    r"z"
-    return t
-
-
-def t_ID(t):
-    r"[a-zA-Z_][a-zA-Z_0-9]*"
-    t.type = "ID"
-    if t.value.lower() in reserved:
-        t.type = t.value.upper()
-    return t
-
-
-def t_NORMSTRING(t):
-    r'"(?:[^"\\]|\\.)*"'
-    return t
-
-
-def t_NORMSTRINGSINGLE(t):
-    r"'(?:[^'\\]|\\.)*'"
-    return t
-
-
-def t_FLOAT(t):
-    r"(\d*\.\d+)|(\d+\.\d*)"
-    t.value = float(t.value)
-    return t
-
-
-def t_INTEGER(t):
-    r"\d+"
-    t.value = int(t.value)
-    return t
-
-
-# Define a rule so we can track line numbers
-def t_newline(t):
-    r"\n+"
-    t.lexer.lineno += len(t.value)
-
-
-# A string containing ignored characters (spaces and tabs)
-t_ignore = " \t"
-
-
-# Error handling rule
-def t_error(t):
-    print("Illegal character '%s'" % t.value[0])
-    t.lexer.skip(1)
-
-
-# Build the lexer
-lexer = lex.lex()
-
-precedence = (
-    ("left", "PLUS", "MINUS"),
-    ("left", "TIMES", "DIVIDE"),
-    ("right", "UMINUS"),
-    ("right", "UNOT"),
-)
-
-
-# --------- Grammar rules ---------
-
-
-def p_expression_logop(p):
-    """
-    expression : expression AND expression
-               | expression OR expression
-    """
-    p[0] = ("logop", p[2], p[1], p[3])
-
-
-def p_expression_unary_not(p):
-    """
-    expression : NOT expression %prec UNOT
-    """
-    p[0] = ("logop", "not", p[2])
-
-
-def p_expression_sameas(p):
-    """
-    expression : SAME FRAGMENT AS expression
-               | SAME molprop_int AS expression
-               | SAME molprop_str AS expression
-    """
-    p[0] = ("sameas", p[2], p[4])
-
-
-def p_expression_within(p):
-    """
-    expression : WITHIN number OF expression
-    """
-    p[0] = ("within", p[2], p[4])
-
-
-def p_expression_exwithin(p):
-    """
-    expression : EXWITHIN number OF expression
-    """
-    p[0] = ("exwithin", p[2], p[4])
-
-
-def p_expression_grouped(p):
-    """
-    expression : LPAREN expression RPAREN
-    """
-    p[0] = ("grouped", p[2])
-
-
-def p_expression_molecule(p):
-    """
-    expression : molecule
-    """
-    p[0] = p[1]
-
-
-def p_molecule(p):
-    """
-    molecule : PROTEIN
-             | NUCLEIC
-             | ION
-             | IONS
-             | LIPID
-             | LIPIDS
-             | WATER
-             | WATERS
-             | BACKBONE
-             | HYDROGEN
-             | NOH
-             | SIDECHAIN
-    """
-    p[0] = ("molecule", p[1])
-
-
-def p_number_expression(p):
-    """
-    expression : number
-    """
-    p[0] = p[1]
-
-
-def p_func_number(p):
-    """
-    number : func
-    """
-    p[0] = p[1]
-
-
-def p_num_funcs(p):
-    """
-    func : ABS LPAREN number RPAREN
-         | SQR LPAREN number RPAREN
-    """
-    p[0] = ("func", p[1], p[3])
-
-
-def p_prop_funcs(p):
-    """
-    func : ABS LPAREN numprop RPAREN
-         | SQR LPAREN numprop RPAREN
-    """
-    p[0] = ("func", p[1], p[3])
-
-
-def p_expression_comp(p):
-    """
-    expression : comp
-    """
-    p[0] = p[1]
-
-
-def p_prop_comp(p):
-    """
-    comp : numprop EQUAL number
-         | numprop LESSER number
-         | numprop GREATER number
-         | numprop LESSEREQUAL number
-         | numprop GREATEREQUAL number
-    """
-    p[0] = ("comp", p[2], p[1], p[3])
-
-
-def p_func_comp(p):
-    """
-    comp : func EQUAL number
-         | func LESSER number
-         | func GREATER number
-         | func LESSEREQUAL number
-         | func GREATEREQUAL number
-    """
-    p[0] = ("comp", p[2], p[1], p[3])
-
-
-def p_number_comp(p):
-    """
-    comp : number EQUAL number
-         | number LESSER number
-         | number GREATER number
-         | number LESSEREQUAL number
-         | number GREATEREQUAL number
-    """
-    p[0] = ("comp", p[2], p[1], p[3])
-
-
-def p_number_mathop(p):
-    """
-    number : number PLUS number
-           | number MINUS number
-           | number TIMES number
-           | number DIVIDE number
-    """
-    p[0] = ("mathop", p[2], p[1], p[3])
-
-
-def p_numprop_mathop(p):
-    """
-    numprop : numprop PLUS number
-            | numprop MINUS number
-            | numprop TIMES number
-            | numprop DIVIDE number
-    """
-    p[0] = ("mathop", p[2], p[1], p[3])
-
-
-def p_expression_molprop(p):
-    """
-    expression : molprop_str_eq
-               | molprop_int_eq
-    """
-    p[0] = p[1]
-
-
-def p_molprop_string_eq(p):
-    """
-    molprop_str_eq : molprop_str string
-    """
-    val2 = p[2]
-    if not isinstance(val2, list):
-        val2 = str(val2)
-    else:
-        val2 = [str(x) for x in val2]
-    p[0] = ("molprop_str_eq", p[1], val2)
-
-
-def p_molprop_string(p):
-    """
-    molprop_str : NAME
-                | ELEMENT
-                | RESNAME
-                | ALTLOC
-                | SEGNAME
-                | SEGID
-                | INSERTION
-                | CHAIN
-    """
-    p[0] = p[1]
-
-
-def p_molprop_int_eq(p):
-    """
-    molprop_int_eq : molprop_int integer
-    """
-    p[0] = ("molprop_int_eq", p[1], p[2])
-
-
-def p_molprop_int(p):
-    """
-    molprop_int : INDEX
-                | SERIAL
-                | RESID
-                | RESIDUE
-    """
-    p[0] = p[1]
-
-
-def p_expression_numprop(p):
-    """
-    expression : numprop
-    """
-    p[0] = p[1]
-
-
-def p_numprop_number(p):
-    """
-    numprop : CHARGE
-            | MASS
-            | OCCUPANCY
-            | XCOOR
-            | YCOOR
-            | ZCOOR
-    """
-    p[0] = ("numprop", p[1])
-
-
-def p_string_list(p):
-    """
-    string : string string
-    """
-    val1 = p[1]
-    val2 = p[2]
-    if not isinstance(val1, list):
-        val1 = [val1]
-    if not isinstance(val2, list):
-        val2 = [val2]
-    p[0] = val1 + val2
-
-
-def p_string(p):
-    """
-    string : ID
-           | NORMSTRING
-           | NORMSTRINGSINGLE
-    """
-    p[0] = p[1]
-
-
-def p_integer_range(p):
-    """
-    integer : integer TO integer
-    """
-    p[0] = list(range(p[1], p[3]))
-
-
-def p_integer_list(p):
-    """
-    integer : integer integer
-    """
-    val1 = p[1]
-    val2 = p[2]
-    if not isinstance(val1, list):
-        val1 = [val1]
-    if not isinstance(val2, list):
-        val2 = [val2]
-    p[0] = val1 + val2
-
-
-def p_number(p):
-    """
-    number : integer
-           | float
-    """
-    p[0] = p[1]
-
-
-def p_number_unary_minus(p):
-    """
-    number : MINUS number %prec UMINUS
-    """
-    p[0] = ("uminus", p[2])
-
-
-def p_integer_unary_minus(p):
-    """
-    integer : MINUS integer %prec UMINUS
-    """
-    p[0] = ("uminus", p[2])
-
-
-def p_float_unary_minus(p):
-    """
-    float : MINUS float %prec UMINUS
-    """
-    p[0] = ("uminus", p[2])
-
-
-def p_integer(p):
-    """
-    integer : INTEGER
-    """
-    p[0] = p[1]
-
-
-def p_float(p):
-    """
-    float : FLOAT
-    """
-    p[0] = p[1]
-
-
-def p_error(p):
-    print(f"Syntax error at {p.value!r}")
-
-
-# Build the parser
-parser = yacc.yacc()
-
-# Parse an expression
-selections = [
-    r"not protein",
-    r"index -15",
-    r"index 1 3 5",
-    r"index 1 to 5",
-    r"name 'A 1'",
-    r"chain X",
-    r"chain 'y'",
-    r"chain 0",
-    r'resname "GL"',
-    r'resname "GL\*"',
-    r"resname ACE NME",
-    r"same fragment as lipid",
-    r"protein and within 8.3 of resname ACE",
-    r"protein and (within -8.3 of resname ACE or exwithin 4 of index 2)",
-    r"mass < 5",
-    r"mass = 4",
-    r"abs(-3)",
-    r"abs(charge)",
-    r"-sqr(charge)",
-    r"abs(charge) > 1",
-    r"abs(charge) <= sqr(4)",
-    r"x < 6",
-    r"x < 6 and x > 3",
-    r"sqr(x-5)+sqr(y+4)+sqr(z) > sqr(5)",
-    "same fragment as resid 5",
-    "same residue as within 8 of resid 100",
-    "same residue as exwithin 8 of resid 100",
-    "same fragment as within 8 of resid 100",
-]
-
-
-def print_first_elems(ast):
-    print(ast[0])
-    for elem in ast:
-        if isinstance(elem, tuple):
-            print_first_elems(elem)
-
-
-for sel in selections:
-    ast = parser.parse(sel)
-    print(f"{sel}:\n   ", ast)
-    # print_first_elems(ast)
-
+from moleculekit.atomselect.languageparser import parser
 from moleculekit.atomselect.analyze import analyze
 from scipy.spatial.distance import cdist
+import numpy as np
+import unittest
 
 molpropmap = {
     "serial": "serial",
@@ -594,7 +64,7 @@ def traverse_ast(mol, analysis, node):
         if molprop in molpropmap:
             return fn(getattr(mol, molpropmap[molprop]), value)
         if molprop == "index":
-            return fn(np.arange(1, mol.numAtoms + 1), value)
+            return fn(np.arange(0, mol.numAtoms), value)
         if molprop == "residue":
             # Unique sequential residue numbering
             return fn(analysis["residues"], value)
@@ -685,46 +155,70 @@ def traverse_ast(mol, analysis, node):
     raise RuntimeError(f"Invalid operation {operation}")
 
 
-mol = Molecule("3ptb")
-mol.serial[10] = -88
-mol.charge[1000:] = -1
+def atomselect(mol, selection, _debug=False, _analysis=None):
+    if _analysis is None:
+        _analysis = analyze(mol)
 
-analysis = analyze(mol)
+    try:
+        ast = parser.parse(selection)
+    except Exception as e:
+        raise RuntimeError(f"Failed to parse selection {selection} with error {e}")
 
-tests = [
-    "serial 1",
-    "serial -88",
-    "index 1",
-    "index 1 2 3",
-    "index 1 to 5",
-    "resname ILE and (index 2)",
-    "resname ALA ILE",
-    "chain A",
-    "charge >= 0",
-    "abs(charge) >= 0",
-    "lipid",
-    "lipids",
-    "ion",
-    "ions",
-    "water",
-    "waters",
-    "noh",
-    "hydrogen",
-    "backbone",
-    "sidechain",
-    "protein",
-    "nucleic",
-    "residue 0",
-    "charge + 5 >= 2+3",
-    "same fragment as resid 17",
-    "same resid as resid 17 18",
-    "same residue as within 8 of resid 100",
-    "same residue as exwithin 8 of resid 100",
-    "same fragment as within 8 of resid 100",
-]
+    if _debug:
+        print(ast)
+    return traverse_ast(mol, _analysis, ast)
 
-for test in tests:
-    ast = parser.parse(test)
-    print(ast)
-    res = traverse_ast(mol, analysis, ast)
-    print(res.sum())
+
+class _TestAtomSelect(unittest.TestCase):
+    def test_atomselect(self):
+        from moleculekit.molecule import Molecule
+
+        selections = [
+            "serial 1",
+            # "serial -88",
+            "index 1",
+            "index 1 2 3",
+            "index 1 to 5",
+            "resname ILE and (index 2)",
+            "resname ALA ILE",
+            "chain A",
+            "charge >= 0",
+            "abs(charge) >= 0",
+            "lipid",
+            "lipids",
+            "ion",
+            "ions",
+            "water",
+            "waters",
+            "noh",
+            "hydrogen",
+            "backbone",
+            "sidechain",
+            "protein",
+            "nucleic",
+            "residue 0",
+            "charge + 5 >= 2+3",
+            "same fragment as resid 17",
+            "same resid as resid 17 18",
+            "same residue as within 8 of resid 100",
+            "same residue as exwithin 8 of resid 100",
+            "same fragment as within 8 of resid 100",
+        ]
+
+        mol = Molecule("3ptb")
+        mol.serial[10] = -88
+        mol.charge[1000:] = -1
+
+        analysis = analyze(mol)
+
+        for sel in selections:
+            with self.subTest(sel=sel):
+                mask1 = atomselect(mol, sel, _analysis=analysis)
+                mask2 = mol.atomselect(sel)
+                assert np.array_equal(
+                    mask1, mask2
+                ), f"{mask1.sum()} vs {mask2.sum()} atoms"
+
+
+if __name__ == "__main__":
+    unittest.main(verbosity=2)
