+++ moleculekit/share/atomselect/atomselect.yaml
@@ -1,0 +1,99 @@
+lipid_resnames:
+  - "DLPE"
+  - "DMPC"
+  - "DPPC"
+  - "GPC"
+  - "LPPC"
+  - "PALM"
+  - "PC"
+  - "PGCL"
+  - "POPC"
+  - "POPE"
+
+ion_resnames:
+  - "AL"
+  - "BA"
+  - "CA"
+  - "CAL"
+  - "CD"
+  - "CES"
+  - "CLA"
+  - "CL"
+  - "CO"
+  - "CS"
+  - "CU"
+  - "CU1"
+  - "CUA"
+  - "HG"
+  - "IN"
+  - "IOD"
+  - "K"
+  - "MG"
+  - "MN3"
+  - "MO3"
+  - "MO4"
+  - "MO5"
+  - "MO6"
+  - "NA"
+  - "NAW"
+  - "OC7"
+  - "PB"
+  - "POT"
+  - "PT"
+  - "RB"
+  - "SOD"
+  - "TB"
+  - "TL"
+  - "WO4"
+  - "YB"
+  - "ZN"
+  - "ZN1"
+  - "ZN2"
+
+protein_backbone_names:
+  - "CA"
+  - "C"
+  - "N"
+  - "O"
+
+protein_terminal_names:
+  - "OT1"
+  - "OT2"
+  - "OXT"
+  - "O1"
+  - "O2"
+
+water_resnames:
+  - "H2O"
+  - "HH0"
+  - "OHH"
+  - "HOH"
+  - "OH2"
+  - "SOL"
+  - "WAT"
+  - "TIP"
+  - "TIP2"
+  - "TIP3"
+  - "TIP4"
+  - "SPC"
+
+nucleic_backbone_names:
+  - "P"
+  - "O1P"
+  - "O2P"
+  - "OP1"
+  - "OP2"
+  - "C3*"
+  - "C3'"
+  - "O3*"
+  - "O3'"
+  - "C4*"
+  - "C4'"
+  - "C5*"
+  - "C5'"
+  - "O5*"
+  - "O5'"
+
+nucleic_terminal_names:
+  - "H5T" 
+  - "H3T"
